Docker setup for IMX8M-EVKB
This setup configures an Docker/Ubuntu environment that can build a
Yocto based OP-TEE / OEMCrypto setup. One can also refer to a repo tree
as mirror/reference outside the Docker environment and the same is true
for DL_DIR (/home/optee/yocto_downloads in the Docker environment) and
SS_STATE (/home/optee/yocto_sstate) in the Docker environment).

One must have access to:
 imx-linux-sumo-imx-4.14.98-2.1.0_8mq_drm-wv-2.2.cand20200235_oemcrypto-v15.tgz

and put that in the same folder as the Dockerfile before trying to build
the Docker image. Paths in Dockerfile and run.sh must be edited so PATHs
matches the environment/computer where one is going to use this.

1. Build the image:
     $ docker build -t imx-widevine

2. Run the image
     $ ./run.sh imx-widevine

3. Further build instructions
     $ cat ~/release_linux_tarball/README_WIDEVINE

   Note that on 'bitbake fsl-image-wpe-drm' it will fail with a compiler
   error in MediaSystem.cpp. Solution to that is:
     $ cd ./tmp/work/aarch64-poky-linux/wpeframework-ocdm-widevine/3.0+gitAUTOINC+052b138536-r0/git/

   Change line 140 in MediaSystem.cpp from:
     if (widevine::Cdm::kSuccess == _cdm->setServiceCertificate((widevine::Cdm::ServiceRole))0, serverCertificate)) {
   to
     if (widevine::Cdm::kSuccess == _cdm->setServiceCertificate((widevine::Cdm::ServiceRole)0, serverCertificate)) {

4. Re-run 'bitbake fsl-image-wpe-drm' and things should work.

5. Continue with the rest of the instructions in the README_WIDEVINE
   file

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>