ubuntu 20:04: add/update the cached version
This version will set up and compile/run OP-TEE on an Ubuntu 20.04
Docker image. It'll leverage cache mechanisms to speed up clone and
build. It will do that by mounting the hosts:
 - repo forest
 - ccache
 - Buildroot download dir

Withing the Docker run-time nothing has to be changed, but for the one
trying this out the "run.sh" script needs to be updated accordingly.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>