Fix env and launch script and
Environment variables in Docker shall be exported using "ENV" and not
"RUN export" as previously done. Also, use full path, so the environment
variables are interpreted correctly from various shells.

This also optimize the launch script, so we don't have to say "Y" when
running repo init and it will also use a jobs flags that is compatible
with the machine where this is been used.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>